mainServer/requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
mysql-connector-python==9.2.0
motor==3.7.1
pymongo==4.13.0
pydantic==2.10.6
pydantic_core==2.27.2
passlib==1.7.4
bcrypt==3.2.0
PyJWT==2.10.1
python-dotenv==1.0.1
python-multipart==0.0.20
redis==6.2.0
python-socketio==5.12.1
python-engineio==4.11.2
email_validator==2.2.0
starlette==0.46.1
typing_extensions==4.12.2